Workspace - throw exception when user tries to copy file to existing destination

+++ desktop-app/src/main/java/pl/jowko/rulerank/desktop/feature/workspace/ContextMenuActions.java
@@ -121,8 +121,13 @@ class ContextMenuActions {
 				
 				if(isCut.get())
 					FileUtils.moveFile(file, newFile);
-				else
+				else {
+					if(newFile.exists()) {
+						throw new FileExistsException("Destination '" + newFile.getAbsolutePath() + "' already exists.");
+					}
 					FileUtils.copyFile(file, newFile, true);
+				}
+				
 				
 			} catch (FileExistsException e) {
 				RuleRankLogger.warn(e.getMessage() + ". Aborting action.");
